OrderItem.java: the relation from OrderItem to MenuItem should be unidirectional ManyToOne, but it was unidirectional OneToOne.
Strangely enough the erroneous OneToOne did not give any errors, when two OrderItems were pointing to the same MenuItem. So JPA/Hibernate doesn't check whether the relation is really OneToOne.

In the database, the unidirectional OneToOne is implemented with a foreign key in OrderItem (which van of course be used as a ManyToOne).

see also: https://forum.hibernate.org/viewtopic.php?f=1&t=1024395&view=previous

RestaurantPopulatorServiceImpl.java:
populateRestaurant() now returns a restaurant with id; that can be handy if you want to do other things with that restaurant. In the current code, the returned restaurant is not used.